Adds factory
