+++ brownie-config.yaml
@@ -1,13 +1,15 @@
 dependencies:
+  - Openzeppelin/openzeppelin-contracts@4.4.2
   - OpenZeppelin/openzeppelin-contracts-upgradeable@4.3.2
 
 autofetch_sources: true
 
 compiler:
   solc:
-    version: null
+    version: 0.8.10
     optimizer:
       enabled: true
-      runs: 10000
+      runs: 9999
     remappings:
-      - '@openzeppelin=OpenZeppelin/openzeppelin-contracts-upgradeable@4.3.2'
+      - "@openzeppelin=Openzeppelin/openzeppelin-contracts@4.4.2"
+      - "@openzeppelin-upgradeable=OpenZeppelin/openzeppelin-contracts-upgradeable@4.3.2"
