+++ interfaces/IVault.sol
@@ -1,11 +1,20 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 pragma solidity ^0.8.10;
 
-import {IERC20Upgradeable as IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20Upgradeable.sol";
+import {IERC20Upgradeable as IERC20} from "@openzeppelin-upgradeable/contracts/token/ERC20/IERC20Upgradeable.sol";
 
 /// @title IVault
 /// @notice Basic MonoVault interface.
+/// @dev This interface should not change frequently and can be used to code interactions
+///      for the users of the Vault. Admin functions are available through the `VaultBase` contract.
 interface IVault is IERC20 {
+    /*///////////////////////////////////////////////////////////////
+                              Vault API Version
+    ///////////////////////////////////////////////////////////////*/
+
+    /// @notice The API version the vault implements
+    function version() external view returns (string memory);
+
     /*///////////////////////////////////////////////////////////////
                               ERC20Detailed
     ///////////////////////////////////////////////////////////////*/
@@ -18,6 +27,44 @@ interface IVault is IERC20 {
 
     /// @notice The Vault shares token decimals.
     function decimals() external view returns (uint8);
+
+    /*///////////////////////////////////////////////////////////////
+                              Batched burns
+    ///////////////////////////////////////////////////////////////*/
+
+    /// @dev Struct for users' batched burning requests.
+    /// @param round Batched burning event index.
+    /// @param shares Shares to burn for the user.
+    struct BatchBurnReceipt {
+        uint256 round;
+        uint256 shares;
+    }
+
+    /// @dev Struct for batched burning events.
+    /// @param totalShares Shares to burn during the event.
+    /// @param amountPerShare Underlying amount per share (this differs from exchangeRate at the moment of batched burning).
+    struct BatchBurn {
+        uint256 totalShares;
+        uint256 amountPerShare;
+    }
+
+    /// @notice Current batched burning round.
+    function batchBurnRound() external view returns (uint256);
+
+    /// @notice Maps user's address to withdrawal request.
+    function userBatchBurnReceipt(address account) external view returns (BatchBurnReceipt memory);
+
+    /// @notice Maps social burning events rounds to batched burn details.
+    function batchBurns(uint256 round) external view returns (BatchBurn memory);
+
+    /// @notice Enter a batched burn event.
+    /// @dev Each user can take part to one batched burn event a time.
+    /// @dev User's shares amount will be staked until the burn happens.
+    /// @param shares Shares to withdraw during the next batched burn event.
+    function enterBatchBurn(uint256 shares) external;
+
+    /// @notice Withdraw underlying redeemed in batched burning events.
+    function exitBatchBurn() external;
 
     /*///////////////////////////////////////////////////////////////
                               ERC4626-like
@@ -40,6 +87,11 @@ interface IVault is IERC20 {
     /// @return underlyingAmount The amount needed to mint `shares` amount of shares.
     function mint(address to, uint256 shares) external returns (uint256);
 
+    /// @notice Returns a user's Vault balance in underlying tokens.
+    /// @param user THe user to get the underlying balance of.
+    /// @return The user's Vault balance in underlying tokens.
+    function balanceOfUnderlying(address user) external view returns (uint256);
+
     /// @notice Calculates the amount of Vault's shares for a given amount of underlying tokens.
     /// @param underlyingAmount The underlying token's amount.
     function calculateShares(uint256 underlyingAmount) external view returns (uint256);
@@ -47,4 +99,20 @@ interface IVault is IERC20 {
     /// @notice Calculates the amount of underlying tokens corresponding to a given amount of Vault's shares.
     /// @param sharesAmount The shares amount.
     function calculateUnderlying(uint256 sharesAmount) external view returns (uint256);
+
+    /// @notice Returns the amount of underlying tokens a share can be redeemed for.
+    /// @return The amount of underlying tokens a share can be redeemed for.
+    function exchangeRate() external view returns (uint256);
+
+    /// @notice Returns the amount of underlying tokens that idly sit in the Vault.
+    /// @return The amount of underlying tokens that sit idly in the Vault.
+    function totalFloat() external view returns (uint256);
+
+    /// @notice Calculate the current amount of locked profit.
+    /// @return The current amount of locked profit.
+    function lockedProfit() external view returns (uint256);
+
+    /// @notice Calculates the total amount of underlying tokens the Vault holds.
+    /// @return The total amount of underlying tokens the Vault holds.
+    function totalUnderlying() external view returns (uint256);
 }
